Remark42: add extra resources capability
This commit introduces an extraResources value which is directly
rendered as part of the chart like a regular Helm template.
This allows users to add some Kubernetes object they might need as part
of their cluster configuration.
I added in the comment an example of my use case because I use a
different type of ingress and it is easier to configure it this way.

Closes: #1250